Create a 'solid_mechanics' directory and move relevant header
