+++ include/solid_mechanics/mf_nh_operator.h
@@ -14,7 +14,7 @@
 #include <deal.II/multigrid/mg_constrained_dofs.h>
 
 #include <base/fe_integrator.h>
-#include <material.h>
+#include <solid_mechanics/material.h>
 
 // Define an operation that takes two tensors $ \mathbf{A} $ and
 // $ \mathbf{B} $ such that their outer-product
