+++ src/version.cc.in
@@ -1,4 +1,4 @@
-#include <version.h>
+#include <base/version.h>
 
 namespace FSI
 {
